requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
elevenlabslib==0.4.2
ffmpeg-python==0.2.0
filelock==3.11.0
flatbuffers==23.3.3
frozenlist==1.3.3
future==0.18.3
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
libclang==16.0.0
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.0.4
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.4
openai-whisper @ git+https://github.com/openai/whisper.git@c09a7ae299c4c34c5839a76380ae407e7d785914
opt-einsum==3.3.0
packaging==23.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.21
pypiwin32==223
pyttsx==1.1
pyttsx3==2.90
pywin32==306
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.0
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tiktoken==0.3.1
torch==2.0.0
tqdm==4.65.0
typing==3.7.4.3
typing_extensions==4.5.0
urllib3==1.26.15
Wave==0.0.2
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
